delete product and products + confirm provider

+++ app/api/[[...route]]/products.ts
@@ -4,6 +4,8 @@ import { Hono } from "hono";
 import { db } from "@/db/drizzle";
 import { products } from "@/db/schema";
 import { and, count, eq, inArray } from "drizzle-orm";
+import { verifyAuth } from "@hono/auth-js";
+import { deleteFromS3 } from "@/lib/s3-upload";
 
 const app = new Hono()
   .get(
@@ -124,6 +126,62 @@ const app = new Hono()
       return c.json({ data });
     }
   )
+  .post(
+    "/delete",
+    verifyAuth(),
+    zValidator(
+      "json",
+      z.object({
+        ids: z.array(z.string()),
+      })
+    ),
+    async (c) => {
+      const auth = c.get("authUser");
+      const values = c.req.valid("json");
+
+      if (!auth || !auth.token?.sub) {
+        return c.json({ error: "Unauthorized" }, 401);
+      }
+
+      const uploadedFiles: string[] = [];
+
+      const existingImages = await db
+        .select({
+          image: products.image,
+        })
+        .from(products)
+        .where(inArray(products.id, values.ids));
+
+      if (!existingImages || existingImages.length === 0) {
+        return c.json({ error: "Failed to fetch images" }, 500);
+      }
+
+      for (const image of existingImages) {
+        if (image.image) {
+          const imageName = image.image.split("/").pop() || "";
+          uploadedFiles.push(imageName);
+        }
+      }
+
+      for (const fileName of uploadedFiles) {
+        const { success, message } = await deleteFromS3(fileName);
+        if (!success) {
+          console.error(`Falha ao deletar ${fileName}: ${message}`);
+          return c.json({ error: message }, 500);
+        }
+      }
+
+      const data = await db
+        .delete(products)
+        .where(inArray(products.id, values.ids));
+
+      if (!data) {
+        return c.json({ error: "Falha ao deletar produtos" }, 500);
+      }
+
+      return c.json({ data });
+    }
+  )
   .delete(
     "/:id",
     zValidator(
@@ -139,10 +197,31 @@ const app = new Hono()
         return c.json({ error: "Missing id" }, 400);
       }
 
+      const [existingImage] = await db
+        .select({
+          image: products.image,
+        })
+        .from(products)
+        .where(eq(products.id, id));
+
+      if (!existingImage || !existingImage.image) {
+        console.error("Failed to fetch image");
+        return c.json({ error: "Falha ao deletar o produto" }, 500);
+      }
+
+      const fileName = existingImage.image.split("/").pop() || "";
+
+      const { success, message } = await deleteFromS3(fileName);
+      if (!success) {
+        console.error(`Falha ao deletar ${fileName}: ${message}`);
+        return c.json({ error: message }, 500);
+      }
+
       const data = await db.delete(products).where(eq(products.id, id));
 
       if (!data) {
-        return c.json({ error: "Failed to delete product" }, 500);
+        console.error("Failed to delete product");
+        return c.json({ error: "Falha ao deletar o produto" }, 500);
       }
 
       return c.json({ data });
